Fix data type for image

+++ OrderService/src/main/java/com/humber/orderservice/entity/Product.java
@@ -10,10 +10,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Product {
-    private int id;
+    private Long id;
     private String name;
     private String description;
     private double price;
     private int stock;
-    private int imageId;
+    private Long imageId;
 }
